requirements.txt
pandas<2.0.0
anndata
scanpy
scikit-learn>=1.2
jupyterlab>=3.0.0
seaborn>=0.11.1
nbconvert>=7.0.0
umap-learn>=0.5.1